% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shRNA_tools.R
\name{circRNASpliceOligoDesigner}
\alias{circRNASpliceOligoDesigner}
\title{Generate shRNAs against circRNA back-splice junctions}
\usage{
circRNASpliceOligoDesigner(
  sp = "dm6",
  input_coordinates = "circ_totest.txt",
  writetab = T,
  output = "shRNAs.tsv",
  shift = T
)
}
\arguments{
\item{sp}{character with Name of the genome to use, available are: dm6, dm3, hg19, hg38, mm10, rn4. Default is dm6}

\item{input_coordinates}{tab separated document with the coordinate of the circRNAs to generate shRNAs in TSV format tab seppatated. It requires a column with Chromosome, Start and End.}

\item{writetab}{a boolean, if writetab=T, then the function will not write any table, instead it will return the table to R}

\item{output}{if writetab = T, it will create the output in this path, default is "shRNAs.tsv"}

\item{shift}{boolean, if shift=T, return also 5´and 3´shift oligos sequences}
}
\value{
if writetab=F, It generates a DataFrame with the original table with the shRNA deisng appended
}
\description{
This script contains ONE funciton that will read a table with circRNA/exon-exon splicing junction coordinates and genenames in the format: Name Chr Start End.
Optional:a column named Strand
So far the following species are available: fly: dm3, dm6 and human: hs19, mice: mm10 and rat: rn4
}
\examples{

#circRNAoligoDesigner(input_coordinates = "../test/circs_totest.txt", writetab=F)


}
