% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shRNA_tools.R
\name{find_rcm}
\alias{find_rcm}
\title{find_rcm Find reverse complementary matches in flanking introns around circRNA}
\usage{
find_rcm(
  sp = "mm39",
  intron_coordinates = "introns.txt",
  blastn = "./",
  ret = F,
  outBlast = "RCM_blast.xls",
  outBlast.text = "RCM_blast.txt"
)
}
\arguments{
\item{sp}{sp character with Name of the genome to use, available are: dm6, dm3, hg19, hg38, mm39, mm10, rn4. Default is mm38}

\item{intron_coordinates}{document with the coordinate of the introns flanking the circRNAs to find the RCMs. It requires a column with Chromosome, Start and End.}

\item{blastn}{is the directory of the blastn script, default "./", in my compute for example is this "/Users/inespatop/Documents/Scripts/blast-2.14.0-h23b05c9_0/bin/blastn"}

\item{ret}{voolean to decide if to return the table as an R object as well as a tab sepparated file}

\item{outBlast}{name of the output file with tab separated statistics from blastn, default is "RCM_blast.xls"}

\item{outBlast.text}{name of the output file to output the alignement grafics, default is "RCM_blast.txt"}
}
\value{
the results of blast
}
\description{
find_rcm Find reverse complementary matches in flanking introns around circRNA
}
\examples{

#circRNAoligoDesigner(input_coordinates = "../test/circs_totest.txt", writetab=F)


}
